+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Exclamation symbols (:exclamation:) note something of importance e.g. breaking c
 - Bot API 5.6 (@TiiFuchs) (#1275)
 - Bot API 5.7 (@TiiFuchs) (#1284)
 - PSR3 (psr/log) version 2 and 3 compatible (@noplanman)
+- Entity implements and uses JsonSerializable now (@TiiFuchs)
 ### Changed
 - Bugfix: Fixed condition in \Longman\TelegramBot\Entities\Chat::isGroupChat() that previously also counted super groups and channels.
 ### Deprecated